Created super collection report page
Added a page to view a report of the super-collection. Made the decision that the Report relation is USELESS, so instead we will implement a page that makes a variety of sql queries to gather data on the users collection, then display that data. Please share thoughts on this decision